+++ src/pyrefdev/config.py
@@ -117,6 +117,7 @@ _packages = [
     Package(pypi="httplib2", index_url="https://httplib2.readthedocs.io/en/latest/"),
     Package(pypi="cyclopts", index_url="https://cyclopts.readthedocs.io/en/latest/"),
     Package(pypi="yib", index_url="https://yib.readthedocs.io/en/latest/"),
+    Package(pypi="passlib", index_url="https://passlib.readthedocs.io/en/stable/"),
     # ENTRY-LINE-MARKER
 ]
 # fmt: on