+++ src/pyrefdev/indexer/add_docs.py
@@ -26,18 +26,8 @@ def add_docs(
         console.fatal(f"Package exists: {package}")
 
     if url is None:
-        url = index.guess_index_url_or_die(package)
-
-    if namespaces:
-        ns_content = ", ".join(f'"{ns}"' for ns in namespaces)
-        ns_str = f", namespaces=[{ns_content}]"
-    else:
-        ns_str = ""
-    config_entry = f'\n    Package(pypi="{package}"{ns_str}, index_url="{url}"),'
-    config_file = Path(config.__file__)
-    config_content = config_file.read_text()
-    config_content = _MARKER.sub(config_entry + r"\g<1>", config_content)
-    config_file.write_text(config_content)
+        url = index.guess_index_url(package, should_die_if_not_found=True)
+    update_config([(package, url, namespaces)])
 
     if crawl:
         importlib.reload(config)
@@ -47,3 +37,22 @@ def add_docs(
             num_threads_per_package=num_threads_per_package,
         )
         update_landing_page_with_packages(config.SUPPORTED_PACKAGES)
+
+
+def update_config(configs: list[tuple[str, str, list[str] | None]]) -> None:
+    config_entries = []
+    for package, url, namespaces in configs:
+        if namespaces:
+            ns_content = ", ".join(f'"{ns}"' for ns in namespaces)
+            ns_str = f", namespaces=[{ns_content}]"
+        else:
+            ns_str = ""
+        config_entries.append(
+            f'    Package(pypi="{package}"{ns_str}, index_url="{url}"),'
+        )
+    config_file = Path(config.__file__)
+    config_content = config_file.read_text()
+    config_content = _MARKER.sub(
+        "\n" + "\n".join(config_entries) + r"\g<1>", config_content
+    )
+    config_file.write_text(config_content)